+++ src/main/java/org/opensearch/searchrelevance/rest/RestGetJudgmentAction.java
@@ -28,17 +28,22 @@ import org.opensearch.rest.BytesRestResponse;
 import org.opensearch.rest.RestRequest;
 import org.opensearch.search.builder.SearchSourceBuilder;
 import org.opensearch.searchrelevance.model.SearchParams;
+import org.opensearch.searchrelevance.settings.SearchRelevanceSettingsAccessor;
 import org.opensearch.searchrelevance.transport.OpenSearchDocRequest;
 import org.opensearch.searchrelevance.transport.judgment.GetJudgmentAction;
 import org.opensearch.searchrelevance.utils.ParserUtils;
 import org.opensearch.transport.client.node.NodeClient;
 
+import lombok.AllArgsConstructor;
+
 /**
  * Rest Action to facilitate requests to get/list judgments.
  */
+@AllArgsConstructor
 public class RestGetJudgmentAction extends BaseRestHandler {
     private static final Logger LOGGER = LogManager.getLogger(RestGetJudgmentAction.class);
     private static final String GET_JUDGMENT_ACTION = "get_judgment_action";
+    private SearchRelevanceSettingsAccessor settingsAccessor;
 
     @Override
     public String getName() {
@@ -52,6 +57,9 @@ public class RestGetJudgmentAction extends BaseRestHandler {
 
     @Override
     protected RestChannelConsumer prepareRequest(RestRequest request, NodeClient client) throws IOException {
+        if (!settingsAccessor.isWorkbenchEnabled()) {
+            return channel -> channel.sendResponse(new BytesRestResponse(RestStatus.FORBIDDEN, "Search Relevance Workbench is disabled"));
+        }
         final String judgmentId = request.param(DOCUMENT_ID);
         // If id is provided, get specific query set
         if (judgmentId != null && !judgmentId.isEmpty()) {