+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Added
 - Added new experiment type for hybrid search ([#10](https://github.com/opensearch-project/search-relevance/pull/26))
+- Added feature flag for search relevance workbench ([#34](https://github.com/opensearch-project/search-relevance/pull/34))
 
 ### Removed